feat(taskslist): Crud relative to the user
With validations and references

+++ app/Http/Controllers/TasksListController.php
@@ -6,12 +6,14 @@ use App\TasksList;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
 
-// TODO: Make auth
 class TasksListController extends Controller
 {
     public function store(Request $request)
     {
-        $validation = Validator::make($request->all(), [
+        $body = $request->all();
+        $user = $request->user();
+        $body['creator'] = $user->id;
+        $validation = Validator::make($body, [
             'name' => 'required|string|max:64',
             'description' => 'string',
             'creator' => 'required|integer|exists:users,id',
@@ -19,36 +21,51 @@ class TasksListController extends Controller
         if($validation->fails()){
             return response()->json($validation->errors(), 417);
         }
-        return TasksList::create($request->all());
+        return TasksList::create($body);
     }
 
-    public function index()
+    public function index(Request $request)
     {
-        return TasksList::all();
+        $user = $request->user();
+        return TasksList::where('creator', $user->id)->get();
     }
 
-    public function show(TasksList $tasksList)
+    public function show(TasksList $tasksList, Request $request)
     {
+        $user = $request->user();
+        if(!$this->hasPermission($tasksList, $user))
+            return response()->json(['message' => 'Permission denied'], 403);
         return $tasksList;
     }
 
     public function update(Request $request, TasksList $tasksList)
     {
-        $validation = Validator::make($request->all(), [
+        $user = $request->user();
+        if(!$this->hasPermission($tasksList, $user))
+            return response()->json(['message' => 'Permission denied'], 403);
+        $body = $request->all();
+        $validation = Validator::make($body, [
             'name' => 'string|max:64',
-            'description' => 'string',
-            'creator' => 'integer|exists:users,id',
+            'description' => 'string'
         ]);
         if($validation->fails()){
             return response()->json($validation->errors(), 417);
         }
-        $tasksList->update($request->all());
+        $tasksList->update($body);
         return $tasksList;
     }
 
-    public function destroy(TasksList $tasksList)
+    public function destroy(TasksList $tasksList, Request $request)
     {
+        $user = $request->user();
+        if(!$this->hasPermission($tasksList, $user))
+            return response()->json(['message' => 'Permission denied'], 403);
         $tasksList->delete();
         return response(null, 200);
+    }
+
+    public function hasPermission($tasksList, $user)
+    {
+        return $tasksList->creator == $user->id;
     }
 }
